helpers: Make assert_file_is_empty actually check file contents
The assert_file_is_empty helper was previously checking the file size to
decide whether a file was empty or not. The issue is that this helper is
currently only used to check /sys/kernel/debug/devices_deferred, which
being a special file exposed by the kernel, always reports a size of 0,
even when it has some content. This resulted in the deferred-probe-empty
test always passing regardless of whether the file was empty or not.

Fix the issue by actually cat'ing the file and checking its contents,
instead of relying on the file size.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>