helpers: Print file contents on assert_file_is_empty
Given that the assert_file_is_empty helper tests for a file's emptiness,
it would be very helpful to know the actual contents of the file when
the test fails. In particular, since this helper is currently only used
by the deferred-probe-empty test, we would be able to know which were
the devices that had their probe deferred when the test fails.

Print the file contents in addition to checking for their presence.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>