helpers: Prevent timeout function from splitting arguments in command
The timeout function receives a command to run as its arguments, but
when it copies them to a different variable with the syntax cmd="$@", it
also joins all arguments into a string, effectively discarding the
quoting of the arguments in the command.

As an example, `timeout $TIMEOUT foo "bar baz"`, which should run
command "foo" with argument "bar baz", instead runs "foo" with two
arguments, "bar" and "baz". Since we want the timeout function to run
the commands exactly as they were passed, use the $@ variable directly
instead of copying it. The only drawback is that this makes the code
slightly less clear.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>